fix(ds-identify): Faster parsing of cloud config
'while read line ... < /file' is very slow with dash, because it
reads character by character.

Change ds-identify to read in the whole configuration file for small
files. The default configuration file is under 4KiB, which happens to
be the block size of modern NVMe storage device.

For large files, fall back to the old read behavior.

Default to 64KiB as the cutoff between small and large, modifiable via
environment variable.

LP: #2030729