feat(opennebula): Use posix shell if bash is unavailable
A hard dependency on bash is undesirable for many reasons. Use 'set'
for getting environment variables, rather than bash features.

Unfortunately, bash implements 'set' in a way that is posix-compliant,
but different from dash/ash[1], so this separate implementations must
for bash and supported posix shells must exist.

Note: The choice to make the datasource configuration implementation a bash
      script was highly unusual. Unfortunately the format is in OpenNebula's
      control, not cloud-init's. Still, adding posix shell support to
      cloud-init's implementation is an incremental improvement.

[1] by outputting ansi-c strings - $'' - rather than multi-line strings